Extract the contexts into a separate workspace crate
