+++ crates/cargo-msrv-cli/src/values/list_msrv_variant.rs
@@ -1,0 +1,7 @@
+use crate::values::{CliValue, CliValues};
+use cargo_msrv_context::types::{DIRECT_DEPS, ListMsrvVariant, ORDERED_BY_MSRV};
+
+pub const VALUES: CliValues<ListMsrvVariant> = CliValues::new(&[
+    CliValue::new(DIRECT_DEPS, ListMsrvVariant::DirectDeps),
+    CliValue::new(ORDERED_BY_MSRV, ListMsrvVariant::OrderedByMSRV),
+]);
