+++ crates/cargo-msrv-context/src/context/show.rs
@@ -1,0 +1,7 @@
+use crate::context::EnvironmentContext;
+
+#[derive(Debug)]
+pub struct ShowContext {
+    /// Resolved environment options
+    pub environment: EnvironmentContext,
+}
