+++ src/compatibility/mod.rs
@@ -1,4 +1,6 @@
+use crate::external_command::cargo_command::CargoCommand;
 use crate::rust::Toolchain;
+use cargo_msrv_context::{CheckCommandContext, FindContext, ToolchainContext, VerifyContext};
 
 mod rustup_toolchain_check;
 #[cfg(test)]
@@ -22,5 +24,36 @@ pub trait IsCompatible {
 
     fn after(&self, _toolchain: &Toolchain) -> TResult<()> {
         Ok(())
+    }
+}
+
+/// The command which is run to determine whether a Rust toolchain is compatible.
+pub trait RunCommandProvider {
+    fn provide_run_command(&self) -> RunCommand;
+}
+
+impl RunCommandProvider for FindContext {
+    fn provide_run_command(&self) -> RunCommand {
+        run_command(&self.check_cmd, &self.toolchain)
+    }
+}
+
+impl RunCommandProvider for VerifyContext {
+    fn provide_run_command(&self) -> RunCommand {
+        run_command(&self.check_cmd, &self.toolchain)
+    }
+}
+
+fn run_command(check_cmd: &CheckCommandContext, toolchain: &ToolchainContext) -> RunCommand {
+    if let Some(custom) = &check_cmd.rustup_command {
+        RunCommand::custom(custom.clone())
+    } else {
+        let cargo_command = CargoCommand::default()
+            .target(Some(toolchain.target))
+            .features(check_cmd.cargo_features.clone())
+            .all_features(check_cmd.cargo_all_features)
+            .no_default_features(check_cmd.cargo_no_default_features);
+
+        RunCommand::from_cargo_command(cargo_command)
     }
 }
