+++ src/error.rs
@@ -1,19 +1,20 @@
 use camino::Utf8PathBuf;
 use owo_colors::OwoColorize;
 use std::env;
-use std::ffi::OsString;
-use std::io;
-use std::path::PathBuf;
 use std::string::FromUtf8Error;
 
-use crate::cli::rust_releases_opts::{ParseEditionError, ParseEditionOrVersionError};
+use crate::cli::rust_releases_opts::ParseEditionOrVersionError;
 use crate::manifest::ManifestParseError;
-use cargo_msrv_cli::types::{
-    ParseListMsrvVariantError, ParseLogLevelError, ParseOutputFormatError, ParseReleaseSourceError,
-    ParseTracingTargetOptionError,
+use cargo_msrv_context::types::{
+    ParseEditionError, ParseListMsrvVariantError, ParseLogLevelError, ParseOutputFormatError,
+    ParseReleaseSourceError, ParseTracingTargetOptionError,
 };
 use cargo_msrv_types::{BareVersion, NoVersionMatchesManifestMsrvError};
 use rust_releases::Release;
+
+pub use cargo_msrv_context::context::error::{
+    Error as ContextError, InvalidUtf8Error, IoError, IoErrorSource, PathError,
+};
 
 use crate::sub_command::{show, verify};
 
@@ -27,8 +28,8 @@ pub enum CargoMSRVError {
     #[error(transparent)]
     CargoMetadata(#[from] cargo_metadata::Error),
 
-    #[error("The default host triple (target) could not be found.")]
-    DefaultHostTripleNotFound,
+    #[error(transparent)]
+    Context(#[from] ContextError),
 
     #[error(transparent)]
     Env(#[from] env::VarError),
@@ -68,9 +69,6 @@ pub enum CargoMSRVError {
 
     #[error(transparent)]
     NoVersionMatchesManifestMSRV(#[from] NoVersionMatchesManifestMsrvError),
-
-    #[error("Unable to find key 'package.rust-version' (or 'package.metadata.msrv') in '{0}'")]
-    NoMSRVKeyInCargoToml(Utf8PathBuf),
 
     #[error(transparent)]
     ParseEdition(#[from] ParseEditionError),
@@ -194,40 +192,6 @@ impl From<ParseTracingTargetOptionError> for CargoMSRVError {
 }
 
 #[derive(Debug, thiserror::Error)]
-#[error("IO error: '{error}'. caused by: '{source}'.")]
-pub struct IoError {
-    pub error: io::Error,
-    pub source: IoErrorSource,
-}
-
-#[derive(Debug, thiserror::Error)]
-pub enum IoErrorSource {
-    #[error("Unable to determine current working directory")]
-    CurrentDir,
-
-    #[error("Unable to open file '{0}'")]
-    OpenFile(Utf8PathBuf),
-
-    #[error("Unable to read file '{0}'")]
-    ReadFile(Utf8PathBuf),
-
-    #[error("Unable to write file '{0}'")]
-    WriteFile(Utf8PathBuf),
-
-    #[error("Unable to remove file '{0}'")]
-    RemoveFile(Utf8PathBuf),
-
-    #[error("Unable to rename file '{0}'")]
-    RenameFile(Utf8PathBuf),
-
-    #[error("Unable to spawn process '{0:?}'")]
-    SpawnProcess(OsString),
-
-    #[error("Unable to collect output from '{0:?}', or process did not terminate properly")]
-    WaitForProcessAndCollectOutput(OsString),
-}
-
-#[derive(Debug, thiserror::Error)]
 pub enum SetMsrvError {
     #[error(
         "Unable to set the MSRV in the 'package.metadata' table: 'package.metadata' is not a table"
@@ -331,48 +295,6 @@ pub struct RustupAddTargetError {
     pub targets: String,
     pub toolchain_spec: String,
     pub stderr: String,
-}
-
-#[derive(Debug, thiserror::Error)]
-pub enum PathError {
-    #[error("'{}' does not exist", .0.display())]
-    DoesNotExist(PathBuf),
-
-    #[error("No parent directory for '{}'", .0.display())]
-    NoParent(PathBuf),
-
-    #[error(transparent)]
-    InvalidUtf8(#[from] InvalidUtf8Error),
-}
-
-#[derive(Debug, thiserror::Error)]
-#[error(transparent)]
-pub struct InvalidUtf8Error {
-    error: Utf8PathErrorInner,
-}
-
-impl From<camino::FromPathError> for InvalidUtf8Error {
-    fn from(value: camino::FromPathError) -> Self {
-        Self {
-            error: Utf8PathErrorInner::FromPath(value),
-        }
-    }
-}
-
-impl From<camino::FromPathBufError> for InvalidUtf8Error {
-    fn from(value: camino::FromPathBufError) -> Self {
-        Self {
-            error: Utf8PathErrorInner::FromPathBuf(value),
-        }
-    }
-}
-
-#[derive(Debug, thiserror::Error)]
-enum Utf8PathErrorInner {
-    #[error("Path contains non UTF-8 characters")]
-    FromPath(camino::FromPathError),
-    #[error("Path contains non UTF-8 characters (path: '{}')", .0.as_path().display())]
-    FromPathBuf(camino::FromPathBufError),
 }
 
 #[derive(Debug, thiserror::Error)]
