+++ src/reporter/event/types/list_result/ordered_by_msrv.rs
@@ -1,11 +1,11 @@
 use super::display_option;
 use super::display_vec;
-use crate::context::list::ORDERED_BY_MSRV;
 use crate::dependency_graph::DependencyGraph;
 use crate::reporter::event::types::list_result::metadata::{format_version, package_msrv};
 use crate::reporter::formatting::table;
 use crate::semver;
 use cargo_metadata::Package;
+use cargo_msrv_context::types::ORDERED_BY_MSRV;
 use petgraph::visit::Bfs;
 use std::collections::BTreeMap;
 use std::fmt;
