+++ src/rust/mod.rs
@@ -1,4 +1,3 @@
-pub(crate) mod default_target;
 mod release;
 pub mod release_index;
 pub(crate) mod releases_filter;
