+++ src/sub_command/verify.rs
@@ -1,18 +1,14 @@
-use camino::Utf8PathBuf;
-use cargo_metadata::MetadataCommand;
-use std::convert::TryFrom;
-
 use rust_releases::{Release, ReleaseIndex};
 
 use crate::compatibility::IsCompatible;
-use crate::context::{EnvironmentContext, VerifyContext};
 use crate::error::{CargoMSRVError, TResult};
-use crate::manifest::CargoManifest;
 use crate::outcome::Compatibility;
 use crate::reporter::Reporter;
 use crate::reporter::event::VerifyResult;
 use crate::rust::Toolchain;
 use crate::sub_command::SubCommand;
+use cargo_msrv_context::VerifyContext;
+use cargo_msrv_context::context::verify::{RustVersion, RustVersionSource};
 use cargo_msrv_types::BareVersion;
 
 /// Verifier which determines whether a given Rust version is deemed compatible or not.
@@ -121,61 +117,11 @@ pub struct VerifyFailed {
 
 impl From<RustVersion> for VerifyFailed {
     fn from(value: RustVersion) -> Self {
+        let (rust_version, source) = value.into_parts();
+
         VerifyFailed {
-            rust_version: value.rust_version,
-            source: value.source,
-        }
-    }
-}
-
-/// A combination of a bare (two- or three component) Rust version and the source which was used to
-/// locate this version.
-#[derive(Clone, Debug)]
-pub struct RustVersion {
-    rust_version: BareVersion,
-    source: RustVersionSource,
-}
-
-impl RustVersion {
-    pub fn from_arg(rust_version: BareVersion) -> Self {
-        Self {
             rust_version,
-            source: RustVersionSource::Arg,
+            source,
         }
     }
-
-    pub fn try_from_environment(env: &EnvironmentContext) -> TResult<Self> {
-        let manifest_path = env.manifest();
-
-        let metadata = MetadataCommand::new()
-            .manifest_path(&manifest_path)
-            .exec()?;
-        CargoManifest::try_from(metadata)?
-            .minimum_rust_version()
-            .ok_or_else(|| CargoMSRVError::NoMSRVKeyInCargoToml(manifest_path.clone()))
-            .map(|v| RustVersion {
-                rust_version: v.clone(),
-                source: RustVersionSource::Manifest(manifest_path.clone()),
-            })
-    }
-
-    /// Get the bare (two- or three component) version specifying the Rust version.
-    pub fn version(&self) -> &BareVersion {
-        &self.rust_version
-    }
-
-    /// Get the version and discard all else.
-    pub fn into_version(self) -> BareVersion {
-        self.rust_version
-    }
-}
-
-/// Source used to obtain a Rust version for the verifier.
-#[derive(Clone, Debug, thiserror::Error)]
-enum RustVersionSource {
-    #[error("as --rust-version argument")]
-    Arg,
-
-    #[error("as MSRV in the Cargo manifest located at '{0}'")]
-    Manifest(Utf8PathBuf),
 }
