+++ tests/common/sub_cmd_verify.rs
@@ -1,6 +1,6 @@
 use crate::common::reporter::EventTestDevice;
 use cargo_msrv::cli::CargoCli;
-use cargo_msrv::compatibility::RustupToolchainCheck;
+use cargo_msrv::compatibility::{RunCommandProvider, RustupToolchainCheck};
 use cargo_msrv::error::CargoMSRVError;
 use cargo_msrv::{Context, SubCommand, Verify};
 use rust_releases::{Release, ReleaseIndex};
@@ -35,7 +35,7 @@ where
         no_check_feedback,
         false, /* Marking unavailable versions as incompatible, which is always false for `verify`  */
         env,
-        verify_ctx.run_command(),
+        verify_ctx.provide_run_command(),
     );
 
     // Determine the MSRV from the index of available releases.
